Backend/fish_ai/requirements.txt
flask
flask-cors
tensorflow
pymysql
numpy
Pillow